# strava/aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Strava API v3",
    "description" : "The [Swagger Playground](https://developers.strava.com/playground) is the easiest way to familiarize yourself with the Strava API by submitting HTTP requests and observing the responses before you write any client code. It will show what a response will look like with different endpoints depending on the authorization scope you receive from your athletes. To use the Playground, go to https://www.strava.com/settings/api and change your “Authorization Callback Domain” to developers.strava.com. Please note, we only support Swagger 2.0. There is a known issue where you can only select one scope at a time. For more information, please check the section “client code” at https://developers.strava.com/docs.",
    "version" : "3.0.0"
  },
  "servers" : [ {
    "url" : "https://www.strava.com/api/v3"
  } ],
  "security" : [ {
    "strava_oauth" : [ "public" ]
  } ],
  "paths" : {
...
    "/segments/{id}" : {
      "get" : {
        "tags" : [ "Segments" ],
        "summary" : "Get Segment",
        "description" : "Returns the specified segment. read_all scope required in order to retrieve athlete-specific segment information, or to retrieve private segments.",
        "operationId" : "getSegmentById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "The identifier of the segment.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Representation of a segment.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DetailedSegment"
                },
                "example" : {
                  "id" : 229781,
                  "resource_state" : 3,
                  "name" : "Hawk Hill",
                  "activity_type" : "Ride",
                  "distance" : 2684.82,
                  "average_grade" : 5.7,
                  "maximum_grade" : 14.2,
                  "elevation_high" : 245.3,
                  "elevation_low" : 92.4,
                  "start_latlng" : [ 37.8331119, -122.4834356 ],
                  "end_latlng" : [ 37.8280722, -122.4981393 ],
                  "climb_category" : 1,
                  "city" : "San Francisco",
                  "state" : "CA",
                  "country" : "United States",
                  "private" : false,
                  "hazardous" : false,
                  "starred" : false,
                  "created_at" : "2009-09-21T20:29:41Z",
                  "updated_at" : "2018-02-15T09:04:18Z",
                  "total_elevation_gain" : 155.733,
                  "map" : {
                    "id" : "s229781",
                    "polyline" : "}g|eFnpqjVl@En@Md@HbAd@d@^h@Xx@VbARjBDh@OPQf@w@d@k@XKXDFPH\\EbGT`AV`@v@|@NTNb@?XOb@cAxAWLuE@eAFMBoAv@eBt@q@b@}@tAeAt@i@dAC`AFZj@dB?~@[h@MbAVn@b@b@\\d@Eh@Qb@_@d@eB|@c@h@WfBK|AMpA?VF\\\\t@f@t@h@j@|@b@hCb@b@XTd@Bl@GtA?jAL`ALp@Tr@RXd@Rx@Pn@^Zh@Tx@Zf@`@FTCzDy@f@Yx@m@n@Op@VJr@",
                    "resource_state" : 3
                  },
                  "effort_count" : 309974,
                  "athlete_count" : 30623,
                  "star_count" : 2428,
                  "athlete_segment_stats" : {
                    "pr_elapsed_time" : 553,
                    "pr_date" : "1993-04-03",
                    "effort_count" : 2
                  }
                }
              }
            }
          },
          "default" : {
            "description" : "Unexpected error.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Fault"
                }
              }
            }
          }
        }
      }
    },
    "/segments/starred" : {
      "get" : {
        "tags" : [ "Segments" ],
        "summary" : "List Starred Segments",
        "description" : "List of the authenticated athlete's starred segments. Private segments are filtered out unless requested by a token with read_all scope.",
        "operationId" : "getLoggedInAthleteStarredSegments",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "description" : "Page number. Defaults to 1.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer"
          }
        }, {
          "name" : "per_page",
          "in" : "query",
          "description" : "Number of items per page. Defaults to 30.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "default" : 30
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "List of the authenticated athlete's starred segments.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/SummarySegment"
                  },
                  "x-content-type" : "application/json"
                },
                "example" : {
                  "id" : 229781,
                  "resource_state" : 3,
                  "name" : "Hawk Hill",
                  "activity_type" : "Ride",
                  "distance" : 2684.82,
                  "average_grade" : 5.7,
                  "maximum_grade" : 14.2,
                  "elevation_high" : 245.3,
                  "elevation_low" : 92.4,
                  "start_latlng" : [ 37.8331119, -122.4834356 ],
                  "end_latlng" : [ 37.8280722, -122.4981393 ],
                  "climb_category" : 1,
                  "city" : "San Francisco",
                  "state" : "CA",
                  "country" : "United States",
                  "private" : false,
                  "hazardous" : false,
                  "starred" : false,
                  "created_at" : "2009-09-21T20:29:41Z",
                  "updated_at" : "2018-02-15T09:04:18Z",
                  "total_elevation_gain" : 155.733,
                  "map" : {
                    "id" : "s229781",
                    "polyline" : "}g|eFnpqjVl@En@Md@HbAd@d@^h@Xx@VbARjBDh@OPQf@w@d@k@XKXDFPH\\EbGT`AV`@v@|@NTNb@?XOb@cAxAWLuE@eAFMBoAv@eBt@q@b@}@tAeAt@i@dAC`AFZj@dB?~@[h@MbAVn@b@b@\\d@Eh@Qb@_@d@eB|@c@h@WfBK|AMpA?VF\\\\t@f@t@h@j@|@b@hCb@b@XTd@Bl@GtA?jAL`ALp@Tr@RXd@Rx@Pn@^Zh@Tx@Zf@`@FTCzDy@f@Yx@m@n@Op@VJr@",
                    "resource_state" : 3
                  },
                  "effort_count" : 309974,
                  "athlete_count" : 30623,
                  "star_count" : 2428,
                  "athlete_segment_stats" : {
                    "pr_elapsed_time" : 553,
                    "pr_date" : "1993-04-03",
                    "effort_count" : 2
                  }
                }
              }
            }
          },
          "default" : {
            "description" : "Unexpected error.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Fault"
                }
              }
            }
          }
        }
      }
    },
...
    "/segments/explore" : {
      "get" : {
        "tags" : [ "Segments" ],
        "summary" : "Explore segments",
        "description" : "Returns the top 10 segments matching a specified query.",
        "operationId" : "exploreSegments",
        "parameters" : [ {
          "name" : "bounds",
          "in" : "query",
          "description" : "The latitude and longitude for two points describing a rectangular boundary for the search: [southwest corner latitutde, southwest corner longitude, northeast corner latitude, northeast corner longitude]",
          "required" : true,
          "style" : "form",
          "explode" : false,
          "schema" : {
            "maxItems" : 4,
            "minItems" : 4,
            "type" : "array",
            "items" : {
              "type" : "number",
              "format" : "float"
            }
          }
        }, {
          "name" : "activity_type",
          "in" : "query",
          "description" : "Desired activity type.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "running", "riding" ]
          }
        }, {
          "name" : "min_cat",
          "in" : "query",
          "description" : "The minimum climbing category.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "maximum" : 5,
            "minimum" : 0,
            "type" : "integer"
          }
        }, {
          "name" : "max_cat",
          "in" : "query",
          "description" : "The maximum climbing category.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "maximum" : 5,
            "minimum" : 0,
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "List of matching segments.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ExplorerResponse"
                },
                "example" : {
                  "segments" : [ {
                    "id" : 229781,
                    "resource_state" : 2,
                    "name" : "Hawk Hill",
                    "climb_category" : 1,
                    "climb_category_desc" : "4",
                    "avg_grade" : 5.7,
                    "start_latlng" : [ 37.8331119, -122.4834356 ],
                    "end_latlng" : [ 37.8280722, -122.4981393 ],
                    "elev_difference" : 152.8,
                    "distance" : 2684.8,
                    "points" : "}g|eFnpqjVl@En@Md@HbAd@d@^h@Xx@VbARjBDh@OPQf@w@d@k@XKXDFPH\\EbGT`AV`@v@|@NTNb@?XOb@cAxAWLuE@eAFMBoAv@eBt@q@b@}@tAeAt@i@dAC`AFZj@dB?~@[h@MbAVn@b@b@\\d@Eh@Qb@_@d@eB|@c@h@WfBK|AMpA?VF\\\\t@f@t@h@j@|@b@hCb@b@XTd@Bl@GtA?jAL`ALp@Tr@RXd@Rx@Pn@^Zh@Tx@Zf@`@FTCzDy@f@Yx@m@n@Op@VJr@",
                    "starred" : false
                  } ]
                }
              }
            }
          },
          "default" : {
            "description" : "Unexpected error.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Fault"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Fault" : {
        "type" : "object",
        "properties" : {
          "errors" : {
            "type" : "array",
            "description" : "The set of specific errors associated with this fault, if any.",
            "items" : {
              "$ref" : "#/components/schemas/Error"
            }
          },
          "message" : {
            "type" : "string",
            "description" : "The message of the fault."
          }
        },
        "description" : "Encapsulates the errors that may be returned from the API."
      },
      "Error" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string",
            "description" : "The code associated with this error."
          },
          "field" : {
            "type" : "string",
            "description" : "The specific field or aspect of the resource associated with this error."
          },
          "resource" : {
            "type" : "string",
            "description" : "The type of resource associated with this error."
          }
        }
      },
...
      "DetailedSegment" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/SummarySegment"
        }, {
          "type" : "object",
          "properties" : {
            "created_at" : {
              "type" : "string",
              "description" : "The time at which the segment was created.",
              "format" : "date-time"
            },
            "updated_at" : {
              "type" : "string",
              "description" : "The time at which the segment was last updated.",
              "format" : "date-time"
            },
            "total_elevation_gain" : {
              "type" : "number",
              "description" : "The segment's total elevation gain.",
              "format" : "float"
            },
            "map" : {
              "$ref" : "#/components/schemas/PolylineMap"
            },
            "effort_count" : {
              "type" : "integer",
              "description" : "The total number of efforts for this segment"
            },
            "athlete_count" : {
              "type" : "integer",
              "description" : "The number of unique athletes who have an effort for this segment"
            },
            "hazardous" : {
              "type" : "boolean",
              "description" : "Whether this segment is considered hazardous"
            },
            "star_count" : {
              "type" : "integer",
              "description" : "The number of stars for this segment"
            }
          }
        } ]
      },
      "SummarySegment" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "description" : "The unique identifier of this segment",
            "format" : "int64"
          },
          "name" : {
            "type" : "string",
            "description" : "The name of this segment"
          },
          "activity_type" : {
            "type" : "string",
            "enum" : [ "Ride", "Run" ]
          },
          "distance" : {
            "type" : "number",
            "description" : "The segment's distance, in meters",
            "format" : "float"
          },
          "average_grade" : {
            "type" : "number",
            "description" : "The segment's average grade, in percents",
            "format" : "float"
          },
          "maximum_grade" : {
            "type" : "number",
            "description" : "The segments's maximum grade, in percents",
            "format" : "float"
          },
          "elevation_high" : {
            "type" : "number",
            "description" : "The segments's highest elevation, in meters",
            "format" : "float"
          },
          "elevation_low" : {
            "type" : "number",
            "description" : "The segments's lowest elevation, in meters",
            "format" : "float"
          },
          "start_latlng" : {
            "$ref" : "#/components/schemas/LatLng"
          },
          "end_latlng" : {
            "$ref" : "#/components/schemas/LatLng"
          },
          "climb_category" : {
            "type" : "integer",
            "description" : "The category of the climb [0, 5]. Higher is harder ie. 5 is Hors catégorie, 0 is uncategorized in climb_category."
          },
          "city" : {
            "type" : "string",
            "description" : "The segments's city."
          },
          "state" : {
            "type" : "string",
            "description" : "The segments's state or geographical region."
          },
          "country" : {
            "type" : "string",
            "description" : "The segment's country."
          },
          "private" : {
            "type" : "boolean",
            "description" : "Whether this segment is private."
          },
          "athlete_pr_effort" : {
            "$ref" : "#/components/schemas/SummarySegmentEffort"
          },
          "athlete_segment_stats" : {
            "$ref" : "#/components/schemas/SummaryPRSegmentEffort"
          }
        },
        "example" : {
          "country" : "country",
          "private" : true,
          "distance" : 6.0274563,
          "average_grade" : 1.4658129,
          "maximum_grade" : 5.962134,
          "climb_category" : 9,
          "city" : "city",
          "elevation_high" : 5.637377,
          "athlete_pr_effort" : {
            "distance" : 7.386282,
            "start_date_local" : "2000-01-23T04:56:07.000+00:00",
            "activity_id" : 2,
            "elapsed_time" : 4,
            "is_kom" : true,
            "id" : 3,
            "start_date" : "2000-01-23T04:56:07.000+00:00"
          },
          "athlete_segment_stats" : {
            "pr_elapsed_time" : 1,
            "pr_date" : "2000-01-23T04:56:07.000+00:00",
            "effort_count" : 1,
            "pr_activity_id" : 1
          },
          "start_latlng" : [ 7.0614014, 7.0614014 ],
          "elevation_low" : 2.302136,
          "end_latlng" : null,
          "activity_type" : "Ride",
          "name" : "name",
          "id" : 0,
          "state" : "state"
        }
      },
      "LatLng" : {
        "maxItems" : 2,
        "minItems" : 2,
        "type" : "array",
        "description" : "A pair of latitude/longitude coordinates, represented as an array of 2 floating point numbers.",
        "items" : {
          "type" : "number",
          "format" : "float"
        }
      },
      "SummarySegmentEffort" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "description" : "The unique identifier of this effort",
            "format" : "int64"
          },
          "activity_id" : {
            "type" : "integer",
            "description" : "The unique identifier of the activity related to this effort",
            "format" : "int64"
          },
          "elapsed_time" : {
            "type" : "integer",
            "description" : "The effort's elapsed time"
          },
          "start_date" : {
            "type" : "string",
            "description" : "The time at which the effort was started.",
            "format" : "date-time"
          },
          "start_date_local" : {
            "type" : "string",
            "description" : "The time at which the effort was started in the local timezone.",
            "format" : "date-time"
          },
          "distance" : {
            "type" : "number",
            "description" : "The effort's distance in meters",
            "format" : "float"
          },
          "is_kom" : {
            "type" : "boolean",
            "description" : "Whether this effort is the current best on the leaderboard"
          }
        },
        "example" : {
          "distance" : 7.386282,
          "start_date_local" : "2000-01-23T04:56:07.000+00:00",
          "activity_id" : 2,
          "elapsed_time" : 4,
          "is_kom" : true,
          "id" : 3,
          "start_date" : "2000-01-23T04:56:07.000+00:00"
        }
      },
      "SummaryPRSegmentEffort" : {
        "type" : "object",
        "properties" : {
          "pr_activity_id" : {
            "type" : "integer",
            "description" : "The unique identifier of the activity related to the PR effort.",
            "format" : "int64"
          },
          "pr_elapsed_time" : {
            "type" : "integer",
            "description" : "The elapsed time ot the PR effort."
          },
          "pr_date" : {
            "type" : "string",
            "description" : "The time at which the PR effort was started.",
            "format" : "date-time"
          },
          "effort_count" : {
            "type" : "integer",
            "description" : "Number of efforts by the authenticated athlete on this segment."
          }
        },
        "example" : {
          "pr_elapsed_time" : 1,
          "pr_date" : "2000-01-23T04:56:07.000+00:00",
          "effort_count" : 1,
          "pr_activity_id" : 1
        }
      },
      "PolylineMap" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "description" : "The identifier of the map"
          },
          "polyline" : {
            "type" : "string",
            "description" : "The polyline of the map, only returned on detailed representation of an object"
          },
          "summary_polyline" : {
            "type" : "string",
            "description" : "The summary polyline of the map"
          }
        },
        "example" : {
          "summary_polyline" : "summary_polyline",
          "id" : "id",
          "polyline" : "polyline"
        }
      },
...
      "ExplorerResponse" : {
        "type" : "object",
        "properties" : {
          "segments" : {
            "type" : "array",
            "description" : "The set of segments matching an explorer request",
            "items" : {
              "$ref" : "#/components/schemas/ExplorerSegment"
            }
          }
        },
        "example" : {
          "segments" : [ {
            "distance" : 2.302136,
            "climb_category" : 3,
            "end_latlng" : null,
            "climb_category_desc" : "NC",
            "name" : "name",
            "id" : 0,
            "elev_difference" : 5.637377,
            "avg_grade" : 1.4658129,
            "start_latlng" : [ 5.962134, 5.962134 ],
            "points" : "points"
          }, {
            "distance" : 2.302136,
            "climb_category" : 3,
            "end_latlng" : null,
            "climb_category_desc" : "NC",
            "name" : "name",
            "id" : 0,
            "elev_difference" : 5.637377,
            "avg_grade" : 1.4658129,
            "start_latlng" : [ 5.962134, 5.962134 ],
            "points" : "points"
          } ]
        }
      },
      "ExplorerSegment" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "description" : "The unique identifier of this segment",
            "format" : "int64"
          },
          "name" : {
            "type" : "string",
            "description" : "The name of this segment"
          },
          "climb_category" : {
            "maximum" : 5,
            "minimum" : 0,
            "type" : "integer",
            "description" : "The category of the climb [0, 5]. Higher is harder ie. 5 is Hors catégorie, 0 is uncategorized in climb_category. If climb_category = 5, climb_category_desc = HC. If climb_category = 2, climb_category_desc = 3."
          },
          "climb_category_desc" : {
            "type" : "string",
            "description" : "The description for the category of the climb",
            "enum" : [ "NC", "4", "3", "2", "1", "HC" ]
          },
          "avg_grade" : {
            "type" : "number",
            "description" : "The segment's average grade, in percents",
            "format" : "float"
          },
          "start_latlng" : {
            "$ref" : "#/components/schemas/LatLng"
          },
          "end_latlng" : {
            "$ref" : "#/components/schemas/LatLng"
          },
          "elev_difference" : {
            "type" : "number",
            "description" : "The segments's evelation difference, in meters",
            "format" : "float"
          },
          "distance" : {
            "type" : "number",
            "description" : "The segment's distance, in meters",
            "format" : "float"
          },
          "points" : {
            "type" : "string",
            "description" : "The polyline of the segment"
          }
        },
        "example" : {
          "distance" : 2.302136,
          "climb_category" : 3,
          "end_latlng" : null,
          "climb_category_desc" : "NC",
          "name" : "name",
          "id" : 0,
          "elev_difference" : 5.637377,
          "avg_grade" : 1.4658129,
          "start_latlng" : [ 5.962134, 5.962134 ],
          "points" : "points"
        }
      },
...
    },
...
    "securitySchemes" : {
      "strava_oauth" : {
        "type" : "oauth2",
        "flows" : {
          "authorizationCode" : {
            "authorizationUrl" : "https://www.strava.com/api/v3/oauth/authorize",
            "tokenUrl" : "https://www.strava.com/api/v3/oauth/token",
            "scopes" : {
              "read" : "Read public segments, public routes, public profile data, public posts, public events, club feeds, and leaderboards",
              "read_all" : "Read private routes, private segments, and private events for the user",
              "profile:read_all" : "Read all profile information even if the user has set their profile visibility to Followers or Only You",
              "profile:write" : "Update the user's weight and Functional Threshold Power (FTP), and access to star or unstar segments on their behalf",
              "activity:read" : "Read the user's activity data for activities that are visible to Everyone and Followers, excluding privacy zone data",
              "activity:read_all" : "The same access as activity:read, plus privacy zone data and access to read the user's activities with visibility set to Only You",
              "activity:write" : "Access to create manual activities and uploads, and access to edit any activities that are visible to the app, based on activity read access level"
            }
          }
        }
      }
    }
  },
...
}